INTYGFV-14815: Refactored resolvePriorityBetweenValidationTypes function and moved it into validationUtils.ts
